// src/app/league/dto/create-league.dto.ts
export class CreateLeagueDto {
  leagueId: string;
  summonerId: string;
  // summonerName: string;
  queueType: string;
  tier: string;
  rank: string;
  leaguePoints: number;
  wins: number;
  losses: number;
  hotStreak: boolean;
  veteran: boolean;
  freshBlood: boolean;
  inactive: boolean;
  region: string;
}